requirements-4.txt
dash==3.2.0
dash_bootstrap_components==2.0.4
matplotlib==3.10.7
numpy==2.3.4
paho_mqtt==2.1.0
pandas==2.3.3
plotly==6.4.0
Requests==2.32.5
seaborn==0.13.2
streamlit==1.51.0
streamlit_autorefresh==1.0.1
requests
